Update member.js
Hey Eric,

This assignment is looking great! Just a few issues with the phone. First, I moved the keyup function outside of the submit. We want that keyup listener to get setup when the document is ready, not after the user has clicked the submit button.

Secondly, I added some code to the phone validation in your onsubmit function. Now, when the phone number is formatted correctly, we rewrite it to the correct format and set it back into the form as the new format.

Keep up the great work!